api: api return type 지정
